Improve data loading
There was no need to fetch the 'out_file' or 'log_file' as it was
already loaded with the report. Improved switching between report and
log/out-file.
Most loaders now keep explicit state of their content to lower the
number of interactions with the backend-server.
Make a start with the new API server based on REST.